+++ SudoSiteReputationIntegrationTests/RealtimeSiteReputationClientIntegrationTests.swift
@@ -9,25 +9,26 @@ import SudoSiteReputation
 
 class RealtimeSiteReputationClientIntegrationTests: IntegrationTestBase {
 
-    var instance: RealtimeSiteReputationClient!
+    var instance: DefaultSudoSiteReputationClient!
+    
     override func setUpWithError() throws {
         try super.setUpWithError()
-        self.instance = try RealtimeSiteReputationClient(userClient: self.userClient)
+        self.instance = try DefaultSudoSiteReputationClient(userClient: self.userClient)
     }
 
     override func tearDownWithError() throws {
         try super.tearDownWithError()
     }
-
-    func test_getReputation_returns_knownGoodSite() async throws {
-        let uri = "http://www.google.com"
+    
+    func test_getReputation_returns_unknownSite() async throws {
+        let uri = "http://www.anonyome.com"
         let result = try await instance.getSiteReputation(url: uri)
-        XCTAssertEqual(result.isMalicious, .notMalicious)
+        XCTAssertEqual(result.status, .unkown)
     }
 
     func test_getReputation_returns_knownBadSite() async throws {
         let uri = "http://malware.wicar.org/data/eicar.com"
         let result = try await instance.getSiteReputation(url: uri)
-        XCTAssertEqual(result.isMalicious, .malicious)
+        XCTAssertEqual(result.status, .malicious)
     }
 }